SQLite MVCC storage engine
This is an initial commit to experimenting with a MVCC storage engine
for SQLite. The plan is to implement a new set of VDBE opcodes to
leverage a MVCC implementation we've been working on:

  https://github.com/penberg/mvcc-rs

We also add a `TF_MVCC` table flag that can be set at table creation
time, which determines which set of opcodes to use.