mvcc: move transaction management to OP_Transaction and OP_Halt
Following how the upper layers expect transactions to be created
and terminated, we also create a transaction during a Transaction
opcode. What remains to be fixed is that the code currently assumes
a single transaction to be in-flight for a single db connection,
which is not necessarily true, either for MVCC transactions
or just read transactions. We need to manage in-flight transactions.