sqlite3: DiskANN vector indexing
This patch adds the DiskANN vector indexing code to the source tree.
It's the lowest level part of the vector search functionality, which
manages a vector index on the disk.

The data structure and algorithm we're following is LM-DiskANN, which is
described in the following paper:

https://cse.unl.edu/~yu/homepage/publications/paper/2023.LM-DiskANN-Low%20Memory%20Footprint%20in%20Disk-Native%20Dynamic%20Graph-Based%20ANN%20Indexing.pdf

What the data structure provides is fast approximate nearest-neighbor
(ANN) search with low memory consumption by having neighbour vectors
stored next to the vector itself in the index pages. When we search the
index, we, therefore, can quickly traverse the nearest neighbours using
best first search (BFS).

For more details on the LM-DiskANN algorithm, please refer to the paper.